Module 4 Coding Assignment Soultion
Souliton for Javascript lecture: looping over an array of names and
printing out either a "hello" or "goodbye" to that name to the browser
console. If the name starts with a letter j or J, console prints out
Goodbye JSomeName. If the name starts with any other letter, console
prints out Hello SomeName.