Module 5 Coding Assignment Solution
The front page of our web app contains 3 clickable tiles: Menu,
Specials, and Map. Clicking Special tile, user will be taken to a single
category page where all the menu items that belong to the Specials menu
category will be shown. The web app takes the user to a random single
category menu page, listing menu items in the category, be it “Lunch”,
“Dinner”, “Sushi”, etc.